refactor: extrair interfaces para um arquivo
