+++ src/app/features/auth/services/auth.service.ts
@@ -1,14 +1,10 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 
-export interface LoginCredentials {
-  email: string;
-  password: string;
-}
-
-export interface LoginResponse {
-  token?: string;
-}
+import {
+  AdminLoginCredentials,
+  AdminLoginResponse,
+} from 'src/app/shared/interfaces/auth.interfaces';
 
 @Injectable({
   providedIn: 'root',
@@ -16,8 +12,8 @@ export interface LoginResponse {
 export class AuthService {
   constructor(private http: HttpClient) {}
 
-  login(credentials: LoginCredentials) {
+  login(credentials: AdminLoginCredentials) {
     // TODO: conectar corretamente com o endpoint da API
-    return this.http.post<LoginResponse>('/api/login', credentials);
+    return this.http.post<AdminLoginResponse>('/api/login', credentials);
   }
 }
