+++ src/app/shared/interfaces/auth.interfaces.ts
@@ -1,0 +1,8 @@
+export interface LoginCredentials {
+  email: string;
+  password: string;
+}
+
+export interface LoginResponse {
+  token?: string;
+}
